requirements.txt
pandas
configparser 
PyVCF
setuptools<58
loguru
numpy
